+++ packages/svelte-undp-components/src/lib/util/index.ts
@@ -1,3 +1,4 @@
 export * from './clean.js';
 export * from './handleEnterKey.js';
 export * from './initTippy.js';
+export * from './getRandomColormap.js';