+++ sites/geohub/svelte.config.js
@@ -1,6 +1,11 @@
 import adapter from '@sveltejs/adapter-node';
 import preprocess from 'svelte-preprocess';
 import { resolve } from 'path';
+import { readFileSync } from 'fs';
+import { fileURLToPath } from 'url';
+const file = fileURLToPath(new URL('package.json', import.meta.url));
+const json = readFileSync(file, 'utf8');
+const pkg = JSON.parse(json);
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -14,6 +19,14 @@ const config = {
 		alias: {
 			$components: resolve('./src/components'),
 			$stores: resolve('./src/stores/index.ts')
+		},
+		version: {
+			name: JSON.stringify({
+				version: pkg.version,
+				license: pkg.license,
+				author: pkg.author,
+				homepage: pkg.homepage
+			})
 		}
 	},